move components into separate directory
