+++ src/components/AppSettings.vue
@@ -28,8 +28,8 @@
 import {
 	AppNavigationSettings,
 } from 'nextcloud-vue'
-import NotesService from './NotesService'
-import store from './store'
+import NotesService from '../NotesService'
+import store from '../store'
 
 export default {
 	name: 'AppSettings',
