+++ src/components/EditorPlain.vue
